+++ src/Pages/Admin/AboutUs.js
@@ -1,14 +1,14 @@
-import NavBarComponent from "../../components/Navbar/Navbar";
+// import NavBarComponent from "../../components/Navbar/Navbar";
 
-function Faq() {
-  return (
-    <div>
-      <NavBarComponent />
-      <div className="d-flex m-0">
-HOLA MUNDO
-      </div>
-    </div>
-  );
-}
+// function Faq() {
+//   return (
+//     <div>
+//       <NavBarComponent />
+//       <div className="d-flex m-0">
+// HOLA MUNDO
+//       </div>
+//     </div>
+//   );
+// }
 
-export default Faq;
+// export default Faq;
