+++ src/Pages/Admin/Productos.js
@@ -1,7 +1,7 @@
 import AdminNavbar from "../../components/Navbar/Navbar";
 import Sidebar from "../Admin/Sidebar/Sidebar"; 
 import "./admin.css"
-import TablaProductos from "../Admin/TablaProductos/TablaProductos";
+import TablaProductos from "../../components/TablaProductos/TablaProductos";
 
 function Products() {
     return (
