+++ src/components/TablaProductos/TablaProductos.js
@@ -1,7 +1,6 @@
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './TablaProductos.css'
 import Form from 'react-bootstrap/Form';
-import "bootstrap/dist/css/bootstrap.min.css";
 import { useState } from 'react';
 import { Button, Modal, InputGroup } from 'react-bootstrap';
 
@@ -15,27 +14,27 @@ function TablaProductos() {
 
     <div className="container-fluid">
       <div className="crud shadow-lg p-3 my-5 bg-dark">
-        <div class="row p-2">
+        <div className="row p-2">
 
-          <div class="col-sm-3 mt-5 mb-4 text-gred">
+          <div className="col-sm-3 mt-5 mb-4 text-gred">
             <div className="search">
-              <form class="form-inline">
-                <input class="form-control mr-sm-2" type="search" placeholder="Buscar Producto" aria-label="search" />
+              <form className="form-inline">
+                <input className="form-control mr-sm-2" type="search" placeholder="Buscar Producto" aria-label="search" />
 
               </form>
             </div>
           </div>
-          <div class="col-sm-3 offset-sm-2 mt-5 mb-4 text-gred text-white"><h4>PRODUCTOS</h4></div>
-          <div class="col-sm-3 offset-sm-1 d-flex py-3 px-0 justify-content-end">
+          <div className="col-sm-3 offset-sm-2 mt-5 mb-4 text-gred text-white"><h4>PRODUCTOS</h4></div>
+          <div className="col-sm-3 offset-sm-1 d-flex py-3 px-0 justify-content-end">
 
             <Button className="btn btn-outline-none btn-success" onClick={handleShow}>
               Añadir Producto
             </Button>
           </div>
         </div>
-        <div class="row justify-content-center">
-          <div class="table-responsive" >
-            <table class="table table-hover table-bordered  bg-light">
+        <div className="row justify-content-center">
+          <div className="table-responsive" >
+            <table className="table table-hover table-bordered  bg-light">
               <thead className='text-center'>
                 <tr>
                   <th className='col-1'>id</th>
@@ -55,9 +54,9 @@ function TablaProductos() {
                   <td>30/01/2023</td>
                   <td>Mujer</td>
                   <td className='d-flex justify-content-evenly'>
-                    <a class="view" title="View" data-toggle="tooltip" style={{ color: "grey" }}><i class="material-icons">&#xE417;</i></a>
-                    <a class="edit" title="Edit" data-toggle="tooltip" style={{ color: "darkgreen" }}><i class="material-icons">&#xE254;</i></a>
-                    <a class="delete" title="Delete" data-toggle="tooltip" style={{ color: "black" }}><i class="material-icons">&#xE872;</i></a>
+                    <a className="view" title="View" data-toggle="tooltip" style={{ color: "grey" }}><i className="material-icons">&#xE417;</i></a>
+                    <a className="edit" title="Edit" data-toggle="tooltip" style={{ color: "darkgreen" }}><i className="material-icons">&#xE254;</i></a>
+                    <a className="delete" title="Delete" data-toggle="tooltip" style={{ color: "black" }}><i className="material-icons">&#xE872;</i></a>
                   </td>
                 </tr>
                 <tr>
@@ -67,9 +66,9 @@ function TablaProductos() {
                   <td>30/01/2023</td>
                   <td>Mujer</td>
                   <td className='d-flex justify-content-evenly'>
-                    <a class="view" title="View" data-toggle="tooltip" style={{ color: "grey" }}><i class="material-icons">&#xE417;</i></a>
-                    <a class="edit" title="Edit" data-toggle="tooltip" style={{ color: "darkgreen" }}><i class="material-icons">&#xE254;</i></a>
-                    <a class="delete" title="Delete" data-toggle="tooltip" style={{ color: "black" }}><i class="material-icons">&#xE872;</i></a>
+                    <a className="view" title="View" data-toggle="tooltip" style={{ color: "grey" }}><i className="material-icons">&#xE417;</i></a>
+                    <a className="edit" title="Edit" data-toggle="tooltip" style={{ color: "darkgreen" }}><i className="material-icons">&#xE254;</i></a>
+                    <a className="delete" title="Delete" data-toggle="tooltip" style={{ color: "black" }}><i className="material-icons">&#xE872;</i></a>
                   </td>
                 </tr>
                 <tr>
@@ -79,9 +78,9 @@ function TablaProductos() {
                   <td>30/01/2023</td>
                   <td>Mujer</td>
                   <td className='d-flex justify-content-evenly'>
-                    <a class="view" title="View" data-toggle="tooltip" style={{ color: "grey" }}><i class="material-icons">&#xE417;</i></a>
-                    <a class="edit" title="Edit" data-toggle="tooltip" style={{ color: "darkgreen" }}><i class="material-icons">&#xE254;</i></a>
-                    <a class="delete" title="Delete" data-toggle="tooltip" style={{ color: "black" }}><i class="material-icons">&#xE872;</i></a>
+                    <a className="view" title="View" data-toggle="tooltip" style={{ color: "grey" }}><i className="material-icons">&#xE417;</i></a>
+                    <a className="edit" title="Edit" data-toggle="tooltip" style={{ color: "darkgreen" }}><i className="material-icons">&#xE254;</i></a>
+                    <a className="delete" title="Delete" data-toggle="tooltip" style={{ color: "black" }}><i className="material-icons">&#xE872;</i></a>
                   </td>
                 </tr>
                 <tr>
@@ -91,9 +90,9 @@ function TablaProductos() {
                   <td>30/01/2023</td>
                   <td>Mujer</td>
                   <td className='d-flex justify-content-evenly'>
-                    <a class="view" title="View" data-toggle="tooltip" style={{ color: "grey" }}><i class="material-icons">&#xE417;</i></a>
-                    <a class="edit" title="Edit" data-toggle="tooltip" style={{ color: "darkgreen" }}><i class="material-icons">&#xE254;</i></a>
-                    <a class="delete" title="Delete" data-toggle="tooltip" style={{ color: "black" }}><i class="material-icons">&#xE872;</i></a>
+                    <a className="view" title="View" data-toggle="tooltip" style={{ color: "grey" }}><i className="material-icons">&#xE417;</i></a>
+                    <a className="edit" title="Edit" data-toggle="tooltip" style={{ color: "darkgreen" }}><i className="material-icons">&#xE254;</i></a>
+                    <a className="delete" title="Delete" data-toggle="tooltip" style={{ color: "black" }}><i className="material-icons">&#xE872;</i></a>
                   </td>
                 </tr>
                 <tr>
@@ -103,9 +102,9 @@ function TablaProductos() {
                   <td>30/01/2023</td>
                   <td>Mujer</td>
                   <td className='d-flex justify-content-evenly'>
-                    <a class="view" title="View" data-toggle="tooltip" style={{ color: "grey" }}><i class="material-icons">&#xE417;</i></a>
-                    <a class="edit" title="Edit" data-toggle="tooltip" style={{ color: "darkgreen" }}><i class="material-icons">&#xE254;</i></a>
-                    <a class="delete" title="Delete" data-toggle="tooltip" style={{ color: "black" }}><i class="material-icons">&#xE872;</i></a>
+                    <a className="view" title="View" data-toggle="tooltip" style={{ color: "grey" }}><i className="material-icons">&#xE417;</i></a>
+                    <a className="edit" title="Edit" data-toggle="tooltip" style={{ color: "darkgreen" }}><i className="material-icons">&#xE254;</i></a>
+                    <a className="delete" title="Delete" data-toggle="tooltip" style={{ color: "black" }}><i className="material-icons">&#xE872;</i></a>
                   </td>
                 </tr>
                 <tr>
@@ -115,9 +114,9 @@ function TablaProductos() {
                   <td>30/01/2023</td>
                   <td>Mujer</td>
                   <td className='d-flex justify-content-evenly'>
-                    <a class="view" title="View" data-toggle="tooltip" style={{ color: "grey" }}><i class="material-icons">&#xE417;</i></a>
-                    <a class="edit" title="Edit" data-toggle="tooltip" style={{ color: "darkgreen" }}><i class="material-icons">&#xE254;</i></a>
-                    <a class="delete" title="Delete" data-toggle="tooltip" style={{ color: "black" }}><i class="material-icons">&#xE872;</i></a>
+                    <a className="view" title="View" data-toggle="tooltip" style={{ color: "grey" }}><i className="material-icons">&#xE417;</i></a>
+                    <a className="edit" title="Edit" data-toggle="tooltip" style={{ color: "darkgreen" }}><i className="material-icons">&#xE254;</i></a>
+                    <a className="delete" title="Delete" data-toggle="tooltip" style={{ color: "black" }}><i className="material-icons">&#xE872;</i></a>
                   </td>
                 </tr>
               </tbody>
@@ -138,21 +137,21 @@ function TablaProductos() {
             </Modal.Header>
             <Modal.Body>
               <form>
-                <div class="form-group">
-                  <input type="text" class="form-control" id="exampleInputEmail1" aria-describedby="Name" placeholder="Nombre" />
+                <div className="form-group">
+                  <input type="text" className="form-control" id="exampleInputEmail1" aria-describedby="Name" placeholder="Nombre" />
                 </div>
-                <div class="form-group mt-3">
-                  <input type="text" class="form-control" id="exampleInputEmail1" aria-describedby="Description" placeholder="Descripción" />
+                <div className="form-group mt-3">
+                  <input type="text" className="form-control" id="exampleInputEmail1" aria-describedby="Description" placeholder="Descripción" />
                 </div>
-                <div class="form-group mt-3">
-                  <input type="number" class="form-control" id="exampleInputEmail1" aria-describedby="Price" placeholder="Precio" />
+                <div className="form-group mt-3">
+                  <input type="number" className="form-control" id="exampleInputEmail1" aria-describedby="Price" placeholder="Precio" />
                 </div>
-                <div class="form-group my-3">
-                  <input type="text" class="form-control" id="exampleInputEmail1" aria-describedby="emailHelp" placeholder="URL" />
+                <div className="form-group my-3">
+                  <input type="text" className="form-control" id="exampleInputEmail1" aria-describedby="emailHelp" placeholder="URL" />
                 </div>
-                <div class="form-group my-3">
+                <div className="form-group my-3">
 
-                  <select class="form-control" id="category_field">
+                  <select className="form-control" id="category_field">
                     <option type="disable">Categoría</option>
                     <option>Hombre</option>
                     <option>Mujer</option>
@@ -171,7 +170,7 @@ function TablaProductos() {
 
                   <Form type="disable" aria-label="Text input with checkbox" className='border-0 text-secondary mx-2'>Publicado</Form>
                 </InputGroup>
-                <button class="custom-btn btn-5"><span>Añadir</span></button>
+                <button className="custom-btn btn-5"><span>Añadir</span></button>
               </form>
             </Modal.Body>
           </Modal>
